requirements.txt
# Countdown Treasure Hunt - Python Dependencies
# Install with: uv pip install -r requirements.txt

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Geospatial Analysis
geopandas>=0.14.0
shapely>=2.0.0
rasterio>=1.3.0
pyproj>=3.6.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Mapping and Visualization
folium>=0.14.0
matplotlib>=3.7.0
contextily>=1.3.0

# Web Requests (for satellite imagery download)
requests>=2.31.0

# Image Processing
scikit-learn>=1.3.0

# Optional: Elevation Data
elevation>=1.1.3

# Optional: EXIF Data
piexif>=1.1.3

# Development
jupyter>=1.0.0
